+++ market-batch/src/main/java/batch/server/alarm/domain/event/ScheduledEvent.java
@@ -1,0 +1,10 @@
+package batch.server.alarm.domain.event;
+
+import java.time.LocalDateTime;
+
+public record ScheduledEvent(
+        Runnable job,
+        String taskId,
+        LocalDateTime date
+) {
+}
